fixed seg fault

+++ src/csv.h
@@ -11,7 +11,7 @@ ps_status_t
 ps_csv_get_column(const char* content,
                   size_t content_len,
                   char** val,
-                  int* val_len,
+                  size_t* val_len,
                   int column);
 
 #endif
